Adds in the workbox welcome msg
