+++ test/workbox-core/static/logger.html
@@ -14,6 +14,13 @@
   <button class="js-error">logger.error()</button>
   <button class="js-group">logger.groupCollapsed()</button>
 
+<script>
+  // Fakery to make it load in the window
+  self.registration = {
+    scope: 'inject-scope',
+  };
+</script>
+
 <script src="../../../../packages/workbox-core/build/browser/workbox-core.dev.js"></script>
 <script>
   const logger = workbox.core._private.logger;
